requirements.txt
Biopython==1.76
argparse==1.4.0